backend/requirements.txt
# FastAPI core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# ORM + Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.6

# Environment variables
python-dotenv==1.0.0

# Security
passlib==1.7.4
bcrypt==3.2.2

# Data validation
pydantic[email]==2.5.3

# HTTP requests
requests==2.31.0

python-jose==3.3.0

google-genai